requirements.txt
pyrogram
pyyaml
TgCrypto